Mengascini-Spina/Sistemi-Digitali-M/requirements.txt
absl-py==0.10.0
ap==0.1.4
app==0.0.1
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
bleach==3.2.1
blinker==1.4
botocore==1.19.13
brotlipy==0.7.0
cachetools==4.2.0
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
Django==3.1.3
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-SocketIO==4.3.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-resources==5.1.0
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mutagen==1.45.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.8.0
psycopg2==2.8.6
ptyprocess==0.6.0
pwn==1.0
pwntools==4.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.9
pyelftools==0.27
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.5
pyzmq==19.0.2
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
ROPGadget==6.3
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
sopaper==0.8
sortedcontainers==2.3.0
soupsieve==2.0.1
spotipy==2.16.1
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-metadata==0.27.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.0.4
tqdm==4.55.2
traitlets==5.0.5
typing-extensions==3.7.4.3
unicorn==1.0.2rc3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
